frontend

+++ client/src/pages/Home.page.tsx
@@ -1,13 +1,10 @@
 import {AppShell, Container, Group, TextInput,Text, Button, Card, Stack, AppShellHeader } from '@mantine/core';
-import { Welcome } from '../components/Welcome/Welcome';
-import { ColorSchemeToggle } from '../components/ColorSchemeToggle/ColorSchemeToggle';
-import LoginButton from '@/components/LoginButton/LoginButton';
 import HomeCard from '@/components/HomeCard';
 import PostCard from '@/components/PostCard';
 import SearchBar from '@/components/SearchBar';
-import mediaImage from '@/0tter.jpg'
+import {Header} from '@/components/Header';
+import {NavBar} from '@/components/NavBar';
 import { useEffect, useState } from 'react';
-
 
 interface Tweet {
   id: string,
@@ -16,6 +13,7 @@ interface Tweet {
 
 export function HomePage() {
   const [data, setData] = useState<Tweet []>([]);
+
 
    useEffect(() => {
     fetchData();
@@ -33,8 +31,13 @@ export function HomePage() {
   };
 
   const handleDelete = async (id: string) => {
+    const token = localStorage.getItem('accessToken');
+  if (!token) {
+    console.error('No access token found');
+    return;
+  }
     const response = await fetch(
-      `http://localhost:8080/removeTweetsById?id=${encodeURIComponent(id)}&accessToken=${encodeURIComponent("OWVtcHhTTEFIdkFsQ1RLOEFQNjBPRU9ETDJ5SHBkNThwOHRRdGhXZmF5YTByOjE3MTQwNzYzNTU1NTg6MToxOmF0OjE")}`,{
+      `http://localhost:8080/removeTweetsById?id=${encodeURIComponent(id)}&accessToken=${encodeURIComponent(token)}`,{
     
       method: 'DELETE',
       headers: {
@@ -48,13 +51,17 @@ export function HomePage() {
     } else {
       console.error('Failed to delete tweet');
     } 
-    const newData = data.filter(item => item.id !== id); //?
-    setData(newData);
+ 
   }
   
   const handlePost = async (tweetText: string) => {
+    const token = localStorage.getItem('accessToken');
+  if (!token) {
+    console.error('No access token found');
+    return;
+  }
     const response = await fetch(
-      `http://localhost:8080/addTweets?accessToken=${encodeURIComponent("OWVtcHhTTEFIdkFsQ1RLOEFQNjBPRU9ETDJ5SHBkNThwOHRRdGhXZmF5YTByOjE3MTQwNzYzNTU1NTg6MToxOmF0OjE")}&Tweets=${encodeURIComponent(tweetText)}`, {
+      `http://localhost:8080/addTweets?accessToken=${encodeURIComponent(token)}&Tweets=${encodeURIComponent(tweetText)}`, {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json'
@@ -67,19 +74,7 @@ export function HomePage() {
     }
   
   }
-  const handleSearch = async (searchTerm: string) => {
-    try {
-        const response = await fetch(`http://localhost:8080/findTweetsById?id=${encodeURIComponent(searchTerm)}`); // Replace with your actual API endpoint
-        const jsonData = await response.json();
-        console.log(jsonData);
-        setData(jsonData);
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      }
-}
-
-  
-
+ 
   return (
     <AppShell
     header={{ height: 60 }}
@@ -98,7 +93,7 @@ export function HomePage() {
     <Container style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '100%' }}>
       <Text style={{ color: 'white', fontWeight: 1000 }}>Twitter Fetch</Text>
       
-      </Container>
+  </Container>
   </AppShell.Header>
   <AppShell.Navbar>
   <Container>
@@ -108,52 +103,49 @@ export function HomePage() {
       <a href="/lookup">Look up</a>
       <a href="/recentsearch">Recent search</a>
       <a href="/timeline">Timelines</a>
+      <a href="/alldata">ALL DATA</a>
+      <a href="/searchAll">search Timelines</a>
+      <a href="/searchPost">search profile</a>
+      
     </Stack>
     </Container>
   </AppShell.Navbar>
     <AppShell.Main>
-    <Stack
+  <Stack
       bg="var(--mantine-color-body)"
-      align="center"
-      gap="sm"
-    >
-      <PostCard 
-        onPost={handlePost}
-      />
-       <SearchBar  onSearch={handleSearch}/>
-        {data && <HomeCard id={data.id} tweet={data.tweets} onDelete={handleSearch}/>
-        
-        }
-                
-     
-      {/* <Welcome /> */}
-      {/* <ColorSchemeToggle /> */}
-      
-      {data.map((item: Tweet) => {
-            // console.log(item.id),
-            // console.log(typeof item.id),
-            // console.log(item.tweets),
-            // console.log(typeof item.tweets),
-            return(
-              <HomeCard
-              key={item.id}
-              id={item.id}
-              tweet={item.tweets}
-              onDelete={handleDelete}
-              /> 
-            )
-              
-      })}
-          
-    
+  align="center"
+  gap="sm"
+>
+  <PostCard 
+    onPost={handlePost}
+  />
    
+ 
+ {data.map((item: Tweet) => {
+        // console.log(item.id),
+        // console.log(typeof item.id),
+        // console.log(item.tweets),
+        // console.log(typeof item.tweets),
+        return(
+          <HomeCard
+          key={item.id}
+          id={item.id}
+          tweet={item.tweets}
+          onDelete={handleDelete}
+          /> 
+        )
+          
+  })} 
+          
+
+    
     </Stack>
     </AppShell.Main>
 
 </AppShell>
   
-  
-   
 
+  
+    
   );
 }
